+++ src/main/java/controller/NoticeList.java
@@ -1,6 +1,7 @@
 package controller;
 
 import java.io.IOException;
+import java.util.ArrayList;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,28 +12,20 @@ import javax.servlet.http.HttpServletResponse;
 import service.NoticeService;
 import vo.Notice;
 
-
-@WebServlet("/NoticeOneController")
-public class NoticeOneController extends HttpServlet {
-
+@WebServlet("/NoticeList")
+public class NoticeList extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-		// 세션 유효성 확인 + 로그인 레벨 확인
 		
+		// 세션 유효성 확인 + 로그인 레벨 확인
 		
 		// 서비스 호출
 		NoticeService noticeService = new NoticeService();
-		int noticeCode = Integer.parseInt(request.getParameter("noticeCode"));
-		System.out.println("noticeCode: "+noticeCode);
+		ArrayList<Notice> noticeList = noticeService.getNoticeList();
 		
-		Notice notice = noticeService.getNoticeOne(noticeCode);
-		
-		// 세션에 저장
-		request.setAttribute("notice", notice);
+		request.setAttribute("noticeList", noticeList);
 		
 		// view 호출
-		request.getRequestDispatcher("/WEB-INF/view/notice/noticeOne.jsp").forward(request, response);
-		
+		request.getRequestDispatcher("/WEB-INF/view/notice/noticeList.jsp").forward(request, response);
 	}
 
 }
