+++ src/main/java/controller/NoticeOne.java
@@ -1,6 +1,7 @@
 package controller;
 
 import java.io.IOException;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -8,28 +9,30 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import service.NoticeService;
+import vo.Notice;
 
 
-@WebServlet("/NoticeRemoveController")
-public class NoticeRemoveController extends HttpServlet {
+@WebServlet("/NoticeOne")
+public class NoticeOne extends HttpServlet {
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
 		// 세션 유효성 확인 + 로그인 레벨 확인
 		
 		
-		// 파라메터 넘겨받기
+		// 서비스 호출
+		NoticeService noticeService = new NoticeService();
 		int noticeCode = Integer.parseInt(request.getParameter("noticeCode"));
 		System.out.println("noticeCode: "+noticeCode);
 		
-		// 서비스 호출
-		NoticeService noticeService = new NoticeService();
-		int row = noticeService.removeNotice(noticeCode);
-		if(row == 0) {
-			System.out.println("삭제 실패");
-			response.sendRedirect(request.getContextPath()+"/NoticeOneController?noticeCode="+noticeCode);
-			return;
-		}
-		System.out.println("삭제 성공");
-		response.sendRedirect(request.getContextPath()+"/NoticeListController");
+		Notice notice = noticeService.getNoticeOne(noticeCode);
+		
+		// 세션에 저장
+		request.setAttribute("notice", notice);
+		
+		// view 호출
+		request.getRequestDispatcher("/WEB-INF/view/notice/noticeOne.jsp").forward(request, response);
+		
 	}
+
 }
