+++ src/main/webapp/WEB-INF/view/notice/modifyNotice.jsp
@@ -9,7 +9,7 @@
 	<body>
 		<h1>MODIFY NOTICE</h1>
 		<!-- 유효성 검사 -->
-		<form action="${pageContext.request.contextPath}/NoticeModifyController" method="post">
+		<form action="${pageContext.request.contextPath}/NoticeModify" method="post">
 			<input type="hidden" name="noticeCode" value="${notice.noticeCode}">
 			<table>
 				<tr>
